Fixed a minor issue

+++ chia/consensus/pos_quality.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from chia.util.ints import uint64
-
 # The actual space in bytes of a plot, is expected_plot_size(k) * UI_ACTUAL_SPACE_CONSTANT_FACTO
 # This is not used in consensus, only for display purposes
 UI_ACTUAL_SPACE_CONSTANT_FACTOR = 0.78
@@ -16,4 +14,4 @@ def expected_plot_size(k: int) -> int:
     is necessary to store the entries in the plot.
     """
 
-    return ((2 * k) + 1) * (2 ** (k - 1))
+    return int(((2 * k) + 1) * (2 ** (k - 1)))
